requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

certifi==2019.3.9
chardet==3.0.4
django-annoying==0.10.4
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django==2.1.5
idna==2.8
numpy==1.16.3
pbr==5.1.3
pipenv-to-requirements==0.7.1
pipenv==2018.11.26
pytz==2019.1
requests==2.21.0
scipy==1.2.1
urllib3==1.24.2
virtualenv-clone==0.5.3
virtualenv==16.5.0
